Modifiying pathes due to new directory structure, implementing modified code with only the mass as network output (new training-, plotting- and converting scripts)
In the modified scripts data.py model.py train_nm.py there has been added a path where files are saved and from where they are loaded again.
In the checkout.sh script there has been added the installation of sklearn and the path for the python2env which is also added in the setup_env.sh.
The new files are containing a new training script with a NN wich has only the mass of the Higgs Boson as an output and a new converting and short plotting script for this model.
Furthermore there is a new variables_massonly.json file which is needed for the conversion of the model into json format.