Adding test data in addition to validation data, changing the evaluation of validation data in the training scrpts train_lbn.py and train_lbn_4jets.py, adding some useful plotting code
The data is now split in three parts, training, validation and test data for all four networks. The evaluation of the training and validation loss is correcte for the scripts train_lbn.py and train_lbn_4jets.py.
Some code for evaluationg and  plotting the networks performances is added: this includes 2d histograms, comparison of the distributions in different mass areas and more.